Added support for Analyzers.  AKA Monkeys.
Now when an updateLogs action happens, the top level analyzerService
(AnalyzerMonkeys at the moment) will be told to analyze the project
if it has  any new commits.  The service uses AnalyzerBookmarks
to keep track of where an analyzer last analyzed a project (which
commit id it analyzed last).  The service grabs all the commit
logs that the analyzer hasn't seen and feeds them to it one
commit at a time.