Went back to having the Analyzers be domain objects.
Well...at least that is now one implementation of the analyzerService.  there
may be others.  this makes creating an analyzer simple...just create a domain
object that extends ProjectAnalyzer and add whatever data fields you want on it.

The analyze method is no args, but the analyzer itself has the commit log, the
project and the previous output from the analyzer (or null if this is the first
commit log).

Testing took a hit here though.  ugh.